feat: Add PLAY_AUDIO command
Introduces a new command type 'PLAY_AUDIO' to allow remote audio playback on the client device.
This includes updates to the CommandType enum, CommandMessage interface, and the Dashboard and MobileScanner views to handle sending and receiving this command. A new 'Ghost' icon is also added to the Dashboard for visual indication.